Change Monster health to i32

+++ src/main.rs
@@ -1,5 +1,5 @@
 struct Monster {
-    health: usize,
+    health: i32,
 }
 
 struct Wizard {}
